Fix provisional support for proxying JS `fetch` via Tor SOCKS
